test(clickhouse): switch DefaultIdentityDeleteRow.date_col to i32
Missed in the Date -> Date32 migration. CI surfaced it as
`SchemaMismatch: attempting to (de)serialize ClickHouse type Date32 as
u16` when `delete_with_default_replica_identity` ran. Match the row
struct to the new schema.

+++ etl-destinations/tests/clickhouse_pipeline.rs
@@ -1553,7 +1553,7 @@ struct DefaultIdentityDeleteRow {
     boolean_col: bool,
     text_col: String,
     varchar_col: String,
-    date_col: u16,
+    date_col: i32,
     timestamp_col: i64,
     timestamptz_col: i64,
     time_col: String,
